Move stop util functions to /util/stops
- Move some reused stop helper functions to utils directory

This removes some dependency cycles between hooks and allows importing some hooks straight from /hooks
