+++ ui/src/hooks/stops/useFilterStops.ts
@@ -1,7 +1,4 @@
-import groupBy from 'lodash/groupBy';
-import maxBy from 'lodash/maxBy';
 import partial from 'lodash/partial';
-import { DateTime } from 'luxon';
 import { useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { ScheduledStopPointDefaultFieldsFragment } from '../../generated/graphql';
@@ -12,6 +9,13 @@ import {
   setStopFilterAction,
 } from '../../redux';
 import { Priority } from '../../types/Priority';
+import {
+  filterHighestPriorityCurrentStops,
+  hasPriority,
+  isCurrentStop,
+  isFutureStop,
+  isPastStop,
+} from '../../utils';
 import { useAppDispatch, useAppSelector } from '../redux';
 import { useObservationDateQueryParam } from '../urlQuery';
 
@@ -38,76 +42,10 @@ export interface FilterItem {
   disabled: boolean;
 }
 
-const isFutureStop = <TStop extends ScheduledStopPointDefaultFieldsFragment>(
-  observationDate: DateTime,
-  stop: TStop,
-) => {
-  // if stop has been valid indefinitely from the start, it can never be a future stop
-  if (!stop.validity_start) {
-    return false;
-  }
-
-  // otherwise its validity has to start after the observation date
-  return stop.validity_start > observationDate;
-};
-
-const isPastStop = <TStop extends ScheduledStopPointDefaultFieldsFragment>(
-  observationDate: DateTime,
-  stop: TStop,
-) => {
-  // if stop is valid indefinitely, it can never be a past stop
-  if (!stop.validity_end) {
-    return false;
-  }
-
-  // otherwise its validity has to end before the observation date
-  return stop.validity_end < observationDate;
-};
-
-const isCurrentStop = <TStop extends ScheduledStopPointDefaultFieldsFragment>(
-  observationDate: DateTime,
-  stop: TStop,
-) => {
-  return (
-    !isPastStop(observationDate, stop) && !isFutureStop(observationDate, stop)
-  );
-};
-
 const mapFilterItemsToFilterFunctions = (filterItems: FilterItem[]) =>
   filterItems
     .filter((item) => item.isActive)
     .map((item) => item.filterFunction);
-
-const hasPriority = <TStop extends ScheduledStopPointDefaultFieldsFragment>(
-  priority: Priority,
-  stop: TStop,
-) => stop.priority === priority;
-
-export const filterHighestPriorityCurrentStops = <
-  TStop extends ScheduledStopPointDefaultFieldsFragment,
->(
-  stops: TStop[],
-  observationDate: DateTime,
-  allowDrafts = false,
-) => {
-  // Get all current stops
-  const currentStops = stops.filter(
-    (stop: TStop) =>
-      isCurrentStop(observationDate, stop) &&
-      (allowDrafts || !hasPriority(Priority.Draft, stop)),
-  );
-
-  // Group stops by label
-  const groupedCurrentStops = groupBy(currentStops, 'label');
-
-  // Pick stop instance with the highest priority for each stop label
-  return (
-    Object.values(groupedCurrentStops)
-      .flatMap((stopInstances) => maxBy(stopInstances, 'priority'))
-      // if for some reason the given group is empty, filter out the undefined values
-      .filter((stop) => !!stop) as TStop[]
-  );
-};
 
 export const useFilterStops = () => {
   const { t } = useTranslation();
